docs: 파일정리, component 폴더 생성해서 component 구분해서 저장, 모듈 수정
